refactor specs to sponsors name file
